Move debug raster tiles to sandbox
