+++ application/src/ext/java/org/opentripplanner/ext/debugrastertiles/MapTile.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.inspector.raster;
+package org.opentripplanner.ext.debugrastertiles;
 
 import org.locationtech.jts.geom.Envelope;
