+++ application/src/ext/java/org/opentripplanner/ext/debugrastertiles/TileRendererManager.java
@@ -1,12 +1,12 @@
-package org.opentripplanner.inspector.raster;
+package org.opentripplanner.ext.debugrastertiles;
 
 import java.awt.image.BufferedImage;
 import java.util.HashMap;
 import java.util.Map;
 import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.util.AffineTransformation;
-import org.opentripplanner.api.resource.GraphInspectorTileResource;
-import org.opentripplanner.inspector.raster.TileRenderer.TileRenderContext;
+import org.opentripplanner.ext.debugrastertiles.TileRenderer.TileRenderContext;
+import org.opentripplanner.ext.debugrastertiles.api.resource.GraphInspectorTileResource;
 import org.opentripplanner.routing.api.request.preference.RoutingPreferences;
 import org.opentripplanner.routing.graph.Graph;
 import org.slf4j.Logger;
