+++ application/src/ext/java/org/opentripplanner/ext/debugrastertiles/WalkSafetyEdgeRenderer.java
@@ -1,10 +1,10 @@
-package org.opentripplanner.inspector.raster;
+package org.opentripplanner.ext.debugrastertiles;
 
 import java.awt.Color;
 import java.util.Optional;
-import org.opentripplanner.inspector.raster.EdgeVertexTileRenderer.EdgeVertexRenderer;
-import org.opentripplanner.inspector.raster.EdgeVertexTileRenderer.EdgeVisualAttributes;
-import org.opentripplanner.inspector.raster.EdgeVertexTileRenderer.VertexVisualAttributes;
+import org.opentripplanner.ext.debugrastertiles.EdgeVertexTileRenderer.EdgeVertexRenderer;
+import org.opentripplanner.ext.debugrastertiles.EdgeVertexTileRenderer.EdgeVisualAttributes;
+import org.opentripplanner.ext.debugrastertiles.EdgeVertexTileRenderer.VertexVisualAttributes;
 import org.opentripplanner.street.model.edge.Edge;
 import org.opentripplanner.street.model.edge.StreetEdge;
 import org.opentripplanner.street.model.vertex.IntersectionVertex;
